+++ src/components/SubTaskList/SubTaskProgressBar/SubTaskProgressBarUtils.js
