+++ src/components/SubTaskList/tests/SubTaskList.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import SubTaskList from './SubTaskList';
+import SubTaskList from '../SubTaskList';
 
 const subTaskListComponent = shallow(<SubTaskList />);
 subTaskListComponent.setState({
